refactor: markdown crate の wasm 成果物を packages/markdown-wasm に一本化
wasm のビルド定義・成果物 (pkg/)・TS ローダーが content-html-backfill と
shuntaka-preview.nvim に重複していたのを、共有 workspace パッケージ
packages/markdown-wasm に集約する。

- packages/markdown-wasm: コミット済み release pkg/ + loadWasm/fetchResources
  ローダー + wasm-pack ビルド定義 (リポジトリ内で唯一)
- content-html-backfill: workspace 依存 (markdown-wasm) に切り替え、自前の
  pkg/ とビルドスクリプト・重複ローダーを削除
- shuntaka-preview.nvim: 相対 import で共有 pkg を参照 (lazy.nvim 導入マシンは
  bun install を実行しないため workspace 名解決に頼らない)
- 鮮度担保: 同一テストスイートをコミット済み pkg と CI で都度 dev ビルドする
  .pkg-fresh の両方に実行 (wasm はビルドマシンの絶対パスを含むためバイト比較
  ではなく挙動一致で検証)。crate 変更時にテストが更新されれば、古い pkg では
  コミット済み側のテストが落ちて再ビルドが強制される
