+++ package.json
@@ -7,6 +7,7 @@
   "author": "",
   "workspaces": [
     "apps/*",
+    "packages/*",
     "tools/*",
     "iac/*",
     "docs"
