+++ tools/content-html-backfill/src/index.ts
@@ -1,54 +1,10 @@
 import { mkdir, writeFile } from 'node:fs/promises';
 import { join } from 'node:path';
 import { program } from 'commander';
+// blog-api の markdown crate の wasm 化 (pkg/ コミット済み) とローダー・フェッチは
+// packages/markdown-wasm に集約している
+import { fetchResources, loadWasm } from 'markdown-wasm';
 import mysql from 'mysql2/promise';
-
-// blog-api の markdown crate を wasm-pack でビルドした成果物（pkg/ は gitignore 済み）。
-// 変換ロジックを API と共有するため、TS 側で再実装せず wasm を呼ぶ。
-// 静的 import にすると pkg 未生成時に type-check が落ちるので動的 import にしている。
-type WasmModule = {
-  collectResourceUrls: (markdown: string) => string[];
-  convertMarkdownWithResources: (markdown: string, resources: Record<string, string>) => string;
-};
-
-async function loadWasm(): Promise<WasmModule> {
-  const pkgUrl = new URL('../pkg/markdown.js', import.meta.url).href;
-  try {
-    const mod = (await import(pkgUrl)) as Record<string, unknown>;
-    const impl = (mod.collectResourceUrls ? mod : mod.default) as WasmModule | undefined;
-    if (!impl?.collectResourceUrls || !impl?.convertMarkdownWithResources) {
-      throw new Error('exports not found');
-    }
-    return impl;
-  } catch (e) {
-    throw new Error(
-      `pkg/markdown.js をロードできない。先に \`bun run build:wasm\` を実行すること (${e instanceof Error ? e.message : String(e)})`,
-    );
-  }
-}
-
-// ureq (markdown crate native 実装) と同じ 5 秒タイムアウト・UA でフェッチする。
-// 失敗した URL はマップに入れない → wasm 側は元の URL をそのまま残すフォールバックに入る
-async function fetchResources(urls: string[], timeoutMs: number): Promise<Record<string, string>> {
-  const resources: Record<string, string> = {};
-  await Promise.all(
-    urls.map(async (url) => {
-      try {
-        const res = await fetch(url, {
-          headers: { 'User-Agent': 'Mozilla/5.0 (compatible; LinkCardBot/1.0)' },
-          signal: AbortSignal.timeout(timeoutMs),
-          redirect: 'follow',
-        });
-        if (res.ok) {
-          resources[url] = await res.text();
-        }
-      } catch {
-        // フェッチ失敗はスキップ（変換側でフォールバック）
-      }
-    }),
-  );
-  return resources;
-}
 
 interface ArticleRow extends mysql.RowDataPacket {
   article_id: string;
