+++ tools/shuntaka-preview.nvim/CLAUDE.md
@@ -4,7 +4,7 @@ This file provides guidance to Claude Code (claude.ai/code) when working with co
 
 ## Project Overview
 
-ブログ (shuntaka.dev) と同じ変換ロジックで Markdown をローカルプレビューする Neovim プラグイン。`apps/blog-api/markdown` crate を wasm-pack で wasm 化した成果物 (`pkg/`) を Bun サーバーから呼び、本番とまったく同じ HTML を生成する。スタイルも `apps/web/src/app/globals.css` をそのまま配信して再現する。
+ブログ (shuntaka.dev) と同じ変換ロジックで Markdown をローカルプレビューする Neovim プラグイン。`apps/blog-api/markdown` crate の wasm 成果物とローダーは `packages/markdown-wasm`（pkg/ コミット済み）に集約されており、Bun サーバーはそれを呼んで本番とまったく同じ HTML を生成する。スタイルも `apps/web/src/app/globals.css` をそのまま配信して再現する。
 
 構成は 3 層:
 
@@ -15,14 +15,14 @@ This file provides guidance to Claude Code (claude.ai/code) when working with co
 ## Commands
 
 ```bash
-# wasm 再ビルド (markdown crate 変更時のみ。pkg/ はコミットする)
-bun run build:wasm
-
-# サーバー単体テスト (コミット済み pkg/ を使用)
+# サーバー単体テスト (markdown-wasm のコミット済み pkg/ を使用)
 bun run test
 
 # サーバーを手動起動 (stdin に NDJSON を流して動作確認する用)
 bun run serve
+
+# wasm 再ビルドは markdown crate 変更時に markdown-wasm 側で行う
+(cd ../../packages/markdown-wasm && bun run build:wasm)
 ```
 
 ## Architecture
@@ -33,7 +33,7 @@ bun run serve
 - `lua/shuntaka-preview/buffer.lua` — TextChanged/CursorMoved の autocmd 登録と vim.uv timer による debounce
 - `src/index.ts` — エントリポイント。Bun.serve (HTTP + WebSocket) と stdin NDJSON ループ。記事一覧クリック (WS open) はパス検証後に stdout 通知 + ファイル内容の即時レンダリング
 - `src/protocol.ts` — Lua↔サーバー↔ブラウザ間メッセージの型定義（唯一の型ソース）
-- `src/markdown.ts` — wasm ロードと 2 パス変換（URL 列挙 → fetch → リソース注入）。成功はプロセス寿命、失敗は 30 秒 TTL でキャッシュ
+- `src/markdown.ts` — markdown-wasm のローダーを使った 2 パス変換（URL 列挙 → fetch → リソース注入）。成功はプロセス寿命、失敗は 30 秒 TTL でキャッシュ
 - `src/articles.ts` — frontmatter の除去/title 抽出（blog-api の ArticleFrontmatter::parse と同じ切り出し方）と記事一覧の列挙
 - `src/template.ts` — shell ページ（記事一覧 + pc/mobile 切り替え + iframe）と view ページ（本番と同じ article-body レイアウト）、globals.css の取り込み（Tailwind ディレクティブのみ除去）
 - `static/client.js` — shell 側。WS 受信を iframe に注入し、記事一覧の描画・ソート・pc/mobile 切り替えを行う
@@ -44,7 +44,7 @@ bun run serve
 ## Design Notes
 
 - wasm の公開 API は `collectResourceUrls` / `convertMarkdownWithResources` の 2 つ（2 パス方式）。wasm 内で同期 HTTP ができないため、fetch は TS 側で行う。content-html-backfill と同じパターン
-- `pkg/` はコミットする方針（content-html-backfill は毎回ビルドするので gitignore、こちらは導入マシンに Rust toolchain を要求しないため）。`build:wasm` が wasm-pack の自動生成する `pkg/.gitignore` を削除する
+- markdown-wasm の参照は workspace 名ではなく相対 import（`../../../packages/markdown-wasm/src/index.js`）。lazy.nvim で入れたマシンでは `bun install` を実行せず node_modules が無いため、名前解決に頼らない形で「bun だけで動く」を維持する
 - lazy.nvim はモノレポのサブディレクトリを直接扱えないため、spec の `config` で require の直前に `vim.opt.rtp:append(plugin.dir .. "/tools/shuntaka-preview.nvim")` する（`init` での追加は初回インストールセッションで間に合わない）。`plugin/` ディレクトリは作らず、コマンド登録は `setup()` で行う
 - プレビューは記事ページのラッパー構造 (`article-content > article-content-wrapper > .prose`) のみ再現。article-header や目次等のサイト chrome、X ポスト埋め込み (react-tweet) は対象外
 - ダークモードは `data-theme` を触らず `prefers-color-scheme` 任せ（本番の未設定時デフォルトと同じ挙動）
