+++ vite.config.ts
@@ -16,7 +16,8 @@ export default defineConfig({
       '**/target',
       '**/next-env.d.ts',
       '**/storybook-static',
-      'tools/shuntaka-preview.nvim/pkg',
+      'packages/markdown-wasm/pkg',
+      'packages/markdown-wasm/.pkg-fresh',
     ],
     options: {
       typeAware: true,
@@ -47,7 +48,8 @@ export default defineConfig({
       '**/next-env.d.ts',
       '**/node_modules/**',
       '**/storybook-static',
-      'tools/shuntaka-preview.nvim/pkg',
+      'packages/markdown-wasm/pkg',
+      'packages/markdown-wasm/.pkg-fresh',
     ],
     singleQuote: true,
   },
